send wrong message after 10th message read
This reproduces the real problem I've found with MongooseIM.
There is indeed a bug in MongooseIM which delivers the same message twice
under some circumstances. This fails given PropEr tests which is
expected. Now PropEr tries to shrink the commands' sequence
and generates wrong sequence when not every arguments from state are
passed to the `wait_fo_msgs_*` function.